Module Mcomp_lang.Ll_table

exception SymbolNotFound of Ast.identifier

exception DuplicateEntry of Ast.identifier

type scope

type var_value

val empty_table : scope

Returns a new empty table

val begin_block : scope -> Ast.identifier -> scope

Returns a new scope with the given id

val end_block : scope -> scope

Deletes the last scope

val add_entry : Ast.identifier -> Llvm.llvalue -> bool -> scope -> scope

Adds a new llvalue with the flag is_arr to the scope

val lookup_is_array : Ast.identifier -> scope -> bool

Returns if the id is a an array as formal parameter

val lookup_val : Ast.identifier -> scope -> Llvm.llvalue

Returns the llval with the given id

val is_function : scope -> bool

Returns if the given scope is a function scope

val get_scope_name : scope -> Ast.identifier

Returns the scope name