Module Mcomp_lang.Location

type lexeme_pos = {

line : int;

start_column : int;

end_column : int;

}

val pp_lexeme_pos : Ppx_deriving_runtime.Format.formatter -> lexeme_pos -> Ppx_deriving_runtime.unit

val show_lexeme_pos : lexeme_pos -> Ppx_deriving_runtime.string

val compare_lexeme_pos : lexeme_pos -> lexeme_pos -> Ppx_deriving_runtime.int

val equal_lexeme_pos : lexeme_pos -> lexeme_pos -> Ppx_deriving_runtime.bool

type code_pos = {

start_line : int;

start_column : int;

end_line : int;

end_column : int;

}

val pp_code_pos : Ppx_deriving_runtime.Format.formatter -> code_pos -> Ppx_deriving_runtime.unit

val show_code_pos : code_pos -> Ppx_deriving_runtime.string

val compare_code_pos : code_pos -> code_pos -> Ppx_deriving_runtime.int

val equal_code_pos : code_pos -> code_pos -> Ppx_deriving_runtime.bool

val to_lexeme_position : Stdlib.Lexing.lexbuf -> lexeme_pos

Returns the given lexing position formatted as a lexeme postion

val to_code_position : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> code_pos

Returns the given lexing position formatted as a code postion

val dummy_code_pos : code_pos

Returns a dummy value of all zeroes as a code position

val dummy_lexeme_pos : lexeme_pos

Returns a dummy value of all zeroes as a lexeme position